bug fix on restructure
