+++ src/components/Add-New/add-new.scss
@@ -1,5 +1,5 @@
-@import "../../colors.scss";
-@import "../../mixins.scss";
+@import "../../styles/colors";
+@import "../../styles/mixins";
 
 .add-new-container {
   position: fixed;
