+++ src/components/LightDarkMode/light-dark-mode.scss
@@ -1,5 +1,5 @@
-@import "../../colors.scss";
-@import "../../mixins.scss";
+@import "../../styles/colors";
+@import "../../styles/mixins";
 
 .main-container-relative {
   width: 70px;
