+++ src/components/Make-Todo/make-todo.scss
@@ -1,5 +1,5 @@
-@import "../../colors.scss";
-@import "../../mixins.scss";
+@import "../../styles/colors";
+@import "../../styles/mixins";
 
 .todo-center {
   max-width: 1200px;
