+++ src/components/Remove-Task/remove-task.scss
@@ -1,5 +1,5 @@
-@import "../../colors.scss";
-@import "../../mixins.scss";
+@import "../../styles/colors";
+@import "../../styles/mixins";
 
 .remove-many {
   display: flex;
