+++ src/components/Search-Todo/search-todo.scss
@@ -1,5 +1,5 @@
-@import "../../colors.scss";
-@import "../../mixins.scss";
+@import "../../styles/colors";
+@import "../../styles/mixins";
 
 .search-area {
   @include flex(row nowrap);
