+++ src/components/User-Account/user-account.scss
@@ -1,5 +1,5 @@
-@import "../../colors.scss";
-@import "../../mixins.scss";
+@import "../../styles/colors";
+@import "../../styles/mixins";
 
 .picture-container {
   @include flex(row nowrap);
